Allow arbitrary numbered hier items (WIP)
The problem with this is we don't know when
the hier elements stop, and we must transition to blocks.
When we run out of "subparagraph", "subsection", etc.?
But the grammar may not obey that, because there may
be hier children underneath the blocks (we could check for that).

Pausing this for now.