Treat QUOTE as a block quote
This swaps the semantics from “we support embedded structure but not inline” to “we support block quotes using embedded structure”. In future we can support inline embedded structures using a similar mechanism to how we do footnotes, which would then be different to this mechanism. Without <block>, we can't differentiate between the two.